requirements.txt
streamlit
pandas
matplotlib
ultralytics
pillow
opencv-python-headless
imageio
numpy
